requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==3.7.1; python_version >= '3.7'
appnope==0.1.3; platform_system == 'Darwin'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
argon2-cffi==23.1.0; python_version >= '3.7'
arrow==1.3.0; python_version >= '3.8'
asttokens==2.4.1
async-lru==2.0.4; python_version >= '3.8'
attrs==23.1.0; python_version >= '3.7'
babel==2.13.1; python_version >= '3.7'
beautifulsoup4==4.12.2; python_version >= '3.6'
bleach==6.1.0; python_version >= '3.8'
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_version >= '3.7'
comm==0.2.0; python_version >= '3.8'
debugpy==1.8.0; python_version >= '3.8'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
distro==1.8.0; python_version >= '3.6'
exceptiongroup==1.2.0; python_version < '3.11'
executing==2.0.1; python_version >= '3.5'
fastjsonschema==2.19.0
fqdn==1.5.1
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.2; python_version >= '3.8'
httpx==0.25.2; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
importlib-metadata==7.0.0; python_version < '3.10'
ipykernel==6.27.1; python_version >= '3.8'
ipython==8.18.1; python_version >= '3.9'
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
json5==0.9.14
jsonpointer==2.4
jsonschema-specifications==2023.11.2; python_version >= '3.8'
jsonschema[format-nongpl]==4.20.0; python_version >= '3.8'
jupyter-client==8.6.0; python_version >= '3.8'
jupyter-core==5.5.0; python_version >= '3.8'
jupyter-events==0.9.0; python_version >= '3.8'
jupyter-lsp==2.2.1; python_version >= '3.8'
jupyter-server-terminals==0.4.4; python_version >= '3.8'
jupyter-server==2.12.0; python_version >= '3.8'
jupyterlab-pygments==0.3.0; python_version >= '3.8'
jupyterlab-server==2.25.2; python_version >= '3.8'
jupyterlab-widgets==3.0.9; python_version >= '3.7'
jupyterlab==4.0.9; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
matplotlib-inline==0.1.6; python_version >= '3.5'
mistune==3.0.2; python_version >= '3.7'
nbclient==0.9.0; python_version >= '3.8'
nbconvert==7.12.0; python_version >= '3.8'
nbformat==5.9.2; python_version >= '3.8'
nest-asyncio==1.5.8; python_version >= '3.5'
notebook-shim==0.2.3; python_version >= '3.7'
notebook==7.0.6
numpy==1.26.2
openai==1.3.7
overrides==7.4.0; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pandocfilters==1.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.3; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32'
platformdirs==4.1.0; python_version >= '3.8'
prometheus-client==0.19.0; python_version >= '3.8'
prompt-toolkit==3.0.41; python_version >= '3.7'
psutil==5.9.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic-core==2.14.5; python_version >= '3.7'
pydantic==2.5.2; python_version >= '3.7'
pygments==2.17.2; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python-json-logger==2.0.7; python_version >= '3.6'
pyyaml==6.0.1; python_version >= '3.6'
pyzmq==25.1.2; python_version >= '3.6'
referencing==0.31.1; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rfc3986-validator==0.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rpds-py==0.13.2; python_version >= '3.8'
send2trash==1.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
stack-data==0.6.3
terminado==0.18.0; python_version >= '3.8'
tinycss2==1.2.1; python_version >= '3.7'
tomli==2.0.1; python_version < '3.11'
tornado==6.4; python_version >= '3.8'
tqdm==4.66.1; python_version >= '3.7'
traitlets==5.14.0; python_version >= '3.8'
types-python-dateutil==2.8.19.14
typing-extensions==4.8.0; python_version >= '3.8'
uri-template==1.3.0
urllib3==2.1.0; python_version >= '3.8'
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0; python_version >= '3.8'
widgetsnbextension==4.0.9; python_version >= '3.7'
zipp==3.17.0; python_version >= '3.8'
